requirements.txt
Django==5.0.6
pillow==10.3.0
crispy-tailwind==1.0.3
djangorestframework==3.15.2
django-environ==0.11.2
psycopg2-binary==2.9.9